Move download function from organization settings to reusable util
